feat(sdds-acore/uikit-compose): Make BaseButton internal
